requirement.txt
Django==1.11.18
django-js-asset==1.1.0
django-polymorphic==2.0.3
django-ranged-response==0.2.0
django-simple-captcha==0.5.10
MySQL-python==1.2.5
Pillow==5.4.1
PyMySQL==0.9.3
pytz==2018.9
six==1.12.0
Unidecode==1.0.23
xlrd==1.2.0
xlwt==1.3.0
